==> 3. Anlysis and Results.Rmd <==
---
title: "3. Analysis and Results"
author: "Emily Raubach"
date: "28 October 2018"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## First steps

Load the necessary packages from the library. 

```{r cars, message=FALSE, warning=FALSE}
library("rjson")
library("RCurl")
library("stringr")
library("dplyr")
library("tm")
library("quanteda")
library("devtools")
library("spacyr")
library("purrr")
library("ggplot2")
library("tidytext")
library('Hmisc')
library('corrplot')
library('ggthemes')
library('grid')

# Install Quanteda dictionaries from GitHub
# devtools::install_github("kbenoit/quanteda.dictionaries")
# devtools::install_github("quanteda/quanteda.corpora")
library("quanteda.corpora")
library("quanteda.dictionaries")
```

Import the combined data file. 

```{R}
# import data
# alltexts_clean <- readRDS("df_clean.rds")

# can also load the dataframe with LSD results
lsd <- readRDS("alltexts_clean_lsd.rds")

# can load file with LSD + POS analysis
pos_df <- readRDS("pos_results.rds")

# can load the dataframe with LIWC results
liwc <- readRDS("alltexts_liwc.rds")

# create a corpus // switch to LIWC or LSD file
#    my_corpus <- corpus(alltexts_clean)
# alternate corpus
# my_corpus <- corpus(alltexts_clean_lsd)

# summary
# head(summary(my_corpus))
```

# Descriptive results

```{r}
# For the summary statistics table

# LSD
# define important variables LSD
vars <- c('Sex','Region','Party.General','Christian','age','seniority',
          'educ','JD_degree','speechcount','statevote_2016','anger','positive','negative',
          'fear','joy','disgust','trust','sadness')

vars_num <- c('age','seniority','speechcount','anger','positive','negative',
              'fear','joy','disgust','trust','sadness')

# descriptive results for LSD variables
# describe(lsd[,vars_num])
summary(lsd[,vars_num])

# standard deviations --> change for each variable
sd(lsd$age)

# LIWC
# define important variables LIWC
vars <- c('Sex','Region','Party.General','Christian','age','seniority',
          'educ','JD_degree','speechcount','statevote_2016','posemo',
          'negemo','Tone','we','WPS','Sixltr','function.','cogproc')

vars_num <- c('posemo','negemo','Tone','we','WPS','Sixltr','function.','cogproc')

# descriptive results for LIWC variables
# describe(liwc[,vars])
summary(liwc[,vars_num])

# standard deviations --> change for each variable
sd(liwc$posemo)
```

# Correlation matrix LSD

```{R}
# select relevant variables
vars_corr <- c('age','seniority','positive','negative')

# create matrix
M <- cor(as.matrix(lsd[,vars_corr]))

# mat : is a matrix of data
# ... : further arguments to pass to the native R cor.test function
cor.mtest <- function(mat, ...) {
    mat <- as.matrix(mat)
    n <- ncol(mat)
    p.mat<- matrix(NA, n, n)
    diag(p.mat) <- 0
    for (i in 1:(n - 1)) {
        for (j in (i + 1):n) {
            tmp <- cor.test(mat[, i], mat[, j], ...)
            p.mat[i, j] <- p.mat[j, i] <- tmp$p.value
        }
    }
  colnames(p.mat) <- rownames(p.mat) <- colnames(mat)
  p.mat
}

# matrix of the p-value of the correlation
p.mat <- cor.mtest(as.matrix(lsd[,vars_corr]))

# define colors
col <- colorRampPalette(c("#BB4444", "#EE9988", "#FFFFFF", "#77AADD", "#4477AA"))

# create correlation plot
corrplot(M, method="color", col=col(200),  
         type="upper", order="hclust", 
         addCoef.col = "black", # Add coefficient of correlation
         tl.col="black", tl.srt=45, #Text label color and rotation
         # Combine with significance
         p.mat = p.mat, sig.level = 0.01, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)
```

# Correlation matrix LIWC
Must run above LSD version first

```{R}
# select relevant variables
vars_corr <- c('age','seniority','posemo','negemo','WPS','Sixltr')

# create matrix
M <- cor(as.matrix(liwc[,vars_corr]))

# matrix of the p-value of the correlation
p.mat <- cor.mtest(as.matrix(liwc[,vars_corr]))

# define colors
col <- colorRampPalette(c("#BB4444", "#EE9988", "#FFFFFF", "#77AADD", "#4477AA"))

# create correlation plot
corrplot(M, method="color", col=col(200),  
         type="upper", order="hclust", 
         addCoef.col = "black", # Add coefficient of correlation
         tl.col="black", tl.srt=45, #Text label color and rotation
         # Combine with significance
         p.mat = p.mat, sig.level = 0.01, insig = "blank", 
         # hide correlation coefficient on the principal diagonal
         diag = FALSE)
```

# 1. Emotion

# Box plots
For both positive and negative emotion from LSD and LIWC

```{r}
library(ggthemes)
library(grid)

# define fill and line colors
fill <- "#4271AE"
line <- "#1F3552"

# --------------------------
# LSD

# Boxplot for both Positive and Negative emotion
df1 <- lsd[,c('Party.General','positive')]
df2 <- lsd[,c('Party.General','negative')]

colnames(df1) <- c('party','value')
colnames(df2) <- c('party','value')

df1$score <- 'Positive'
df2$score <- 'Negative'

# combine two dataframes
df <- rbind(df1, df2)

# define as factor
df$score <- factor(df$score)

# create plot
p11 <- ggplot(df, aes(x = party, y = value)) +
        geom_boxplot(fill = fill, colour = line, alpha = 0.7, 
                     width = 0.6, position = position_dodge2(preserve = "total")) +
        scale_y_continuous(name = "Percentage of usage",
                           breaks = seq(0, 15, 2),
                           limits=c(0, 15)) +
        scale_x_discrete(name = "Party") +
        ggtitle("") +
        theme_bw() +
        theme(plot.title = element_text(size = 14, family = "Tahoma", face = "bold"),
              text = element_text(size = 12, family = "Tahoma"),
              axis.title = element_text(face="bold"),
              axis.text.x=element_text(size = 11)) +
        facet_grid(. ~ score) +
        coord_flip()
  
# create plot with differing scales
p12 <- ggplot(df, aes(x = party, y = value)) +
        geom_boxplot(fill = fill, colour = line,
                     alpha = 0.7,
                     width = 0.6, position = position_dodge2(preserve = "total")) +
        scale_y_continuous(name = "Percentage of usage") +
        scale_x_discrete(name = "Party") +
        ggtitle("") +
        theme_bw() +
        theme(plot.title = element_text(size = 14, family = "Tahoma", face = "bold"),
              text = element_text(size = 12, family = "Tahoma"),
              axis.title = element_text(face="bold"),
              axis.text.x=element_text(size = 11)) +
        facet_grid(. ~ score, scales = "free") +
        coord_flip()

# --------------------------
# LIWC

# Boxplot for both Positive and Negative emotion
df1 <- liwc[,c('Party.General','posemo')]
df2 <- liwc[,c('Party.General','negemo')]

colnames(df1) <- c('party','value')
colnames(df2) <- c('party','value')

df1$score <- 'positive'
df2$score <- 'negative'

# combine two dataframes
df <- rbind(df1, df2)

# define as factor
df$score <- factor(df$score)

# create plot
p12 <- ggplot(df, aes(x = party, y = value)) +
        geom_boxplot(fill = fill, colour = line,
                     alpha = 0.7) +
        scale_y_continuous(name = "Score",
                           breaks = seq(0, 15, 2),
                           limits=c(0, 10)) +
        scale_x_discrete(name = "Party") +
        ggtitle("Boxplot of positive and negative emotion by party") +
        theme_bw() +
        theme(plot.title = element_text(size = 14, family = "Tahoma", face = "bold"),
              text = element_text(size = 12, family = "Tahoma"),
              axis.title = element_text(face="bold"),
              axis.text.x=element_text(size = 11)) +
        facet_grid(. ~ score)
```

______________________________________________________________________________________________________

# 1.1. Positive Emotion
______________________________________________________________________________________________________

# t-test results

```{r}
# --------------------------
# LSD

# t-test of two independent samples (Democrats vs. Republcans)
t.test(lsd$positive ~ lsd$Party.General, alternative = 'two.sided', conf.level = 0.95)

# Quality check - no need to report this
# QQ plots for Democrats and Republicans
qqnorm(lsd$positive[lsd$Party.General == 'Democrat'], main = 'Democrats')
qqline(lsd$positive[lsd$Party.General == 'Democrat'])

qqnorm(lsd$positive[lsd$Party.General == 'Republican'], main = 'Republicans')
qqline(lsd$positive[lsd$Party.General == 'Republican'])

# independent 2-group Mann-Whitney U Test 
wilcox.test(lsd$postive ~ lsd$Party.General)

# --------------------------
# LIWC
# t-test of two independent samples (Democrats vs. Republcans)
t.test(liwc$posemo ~ liwc$Party.General, alternative = 'two.sided', conf.level = 0.95)

# Quality check - no need to report this
# QQ plots for Democrats and Republicans
qqnorm(liwc$posemo[liwc$Party.General == 'Democrat'], main = 'Democrats')
qqline(liwc$posemo[liwc$Party.General == 'Democrat'])

qqnorm(liwc$posemo[liwc$Party.General == 'Republican'], main = 'Republicans')
qqline(liwc$posemo[liwc$Party.General == 'Republican'])

# independent 2-group Mann-Whitney U Test 
wilcox.test(liwc$posemo ~ liwc$Party.General)
```

# Regression analysis

```{R}
# Multiple Linear Regression for Positive Emotion

# Model 1 with full sample
fit1 <- lm(positive ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = lsd[lsd$speechcount > 0, ])

# Model 2 with smaller sample based on speech count (more than 9 speeches given)
fit2 <- lm(positive ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = lsd[lsd$speechcount > 9, ])

# Model 3 of interaction effects with party and sex with full sample 
fit3 <- lm(positive ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + Party.General:Sex, # interaction term 
          data = lsd[lsd$speechcount > 0, ])

# Model 4 with addition of the statevote variable and a smaller sample
fit4 <- lm(positive ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + statevote_2016, 
          data = lsd[lsd$speechcount > 0, ])

# show results
summary(fit1)
summary(fit2)
summary(fit3)
summary(fit4)
```

# Diagnostic checks of the regression models

```{R}
# Diagnostic plots (QQ, et al.)
layout(matrix(c(1,2,3,4), 2, 2))
layout(matrix(c(1,2), 2, 1))
plot(fit1)


# Boxplot for Positive affect
boxplot(positive ~ Party.General,data = lsd[alltexts_clean$speechcount > 9,], 
        main="Positive Affect", 
        xlab = "Party", ylab = "Positive Affect")
```

______________________________________________________________________________________________________

# 1.2. Negative Emotion
______________________________________________________________________________________________________

# Box plots
See the script for positive (as they show both variables + LIWC results)

# t-test results

```{r}
# --------------------------
# LSD

# t-test of two independent samples (Democrats vs. Republcans)
t.test(lsd$negative ~ lsd$Party.General, alternative = 'two.sided', conf.level = 0.95)

# Quality check - no need to report this
# QQ plots for Democrats and Republicans
qqnorm(lsd$negative[lsd$Party.General == 'Democrat'], main = 'Democrats')
qqline(lsd$negative[lsd$Party.General == 'Democrat'])

qqnorm(lsd$negative[lsd$Party.General == 'Republican'], main = 'Republicans')
qqline(lsd$negative[lsd$Party.General == 'Republican'])

# independent 2-group Mann-Whitney U Test 
wilcox.test(lsd$negative ~ lsd$Party.General)

# --------------------------
# LIWC
# t-test of two independent samples (Democrats vs. Republcans)
t.test(liwc$negemo ~ liwc$Party.General, alternative = 'two.sided', conf.level = 0.95)

# Quality check - no need to report this
# QQ plots for Democrats and Republicans
qqnorm(liwc$negemo[liwc$Party.General == 'Democrat'], main = 'Democrats')
qqline(liwc$negemo[liwc$Party.General == 'Democrat'])

qqnorm(liwc$negemo[liwc$Party.General == 'Republican'], main = 'Republicans')
qqline(liwc$negemo[liwc$Party.General == 'Republican'])

# independent 2-group Mann-Whitney U Test 
wilcox.test(liwc$negemo ~ liwc$Party.General)

```

# Regression analysis

```{R}
# Multiple Linear Regression for Negative Emotion

# Model 1 with full sample
fit1 <- lm(negative ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = lsd[lsd$speechcount > 0, ])

# Model 2 with smaller sample based on speech count (more than 9 speeches given)
fit2 <- lm(negative ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = lsd[lsd$speechcount > 9, ])

# Model 3 of interaction effects with party and sex with full sample 
fit3 <- lm(negative ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + Party.General:Sex, # interaction term 
          data = lsd[lsd$speechcount > 0, ])

# Model 4 with addition of the statevote variable and a smaller sample
fit4 <- lm(negative ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + statevote_2016, 
          data = lsd[lsd$speechcount > 0, ])

# show results
summary(fit1)
summary(fit2)
summary(fit3)
summary(fit4)
```

# Diagnostic checks of the regression models

```{R}
# Diagnostic plots (QQ, et al.)
layout(matrix(c(1,2,3,4), 2, 2))
plot(fit2)

# Boxplot for negative affect
layout(matrix(c(1,1), 2, 2))
boxplot(negative ~ Party.General,data = lsd[lsd$speechcount > 9,], 
        main="Negative Affect", 
        xlab = "Party", ylab = "Negative Affect")
```



______________________________________________________________________________________________________

# 2. Complexity
______________________________________________________________________________________________________

# Boxplots for BW and WPS


```{r}
# define fill and line colors
fill <- "#4271AE"
line <- "#1F3552"

# Boxplot for big words and words per sentence
df1 <- liwc[,c('Party.General','Sixltr')]
df2 <- liwc[,c('Party.General','WPS')]

colnames(df1) <- c('party','value')
colnames(df2) <- c('party','value')

df1$score <- 'Big words'
df2$score <- 'Mean sentence length'

# combine two dataframes
df <- rbind(df1, df2)

# define as factor
df$score <- factor(df$score)

# create plot
p12 <- ggplot(df, aes(x = party, y = value)) +
        geom_boxplot(fill = fill, colour = line, alpha = 0.7, 
                     width = 0.6, position = position_dodge2(preserve = "total")) +
        scale_y_continuous(name = "",
                           breaks = seq(10, 36, 4),
                           limits=c(10, 36)) +
        scale_x_discrete(name = "Party") +
        ggtitle("") +
        theme_bw() +
        theme(plot.title = element_text(size = 14, family = "Tahoma", face = "bold"),
              text = element_text(size = 12, family = "Tahoma"),
              axis.title = element_text(face="bold"),
              axis.text.x=element_text(size = 11)) +
        facet_grid(. ~ score) +
        coord_flip()


```
______________________________________________________________________________________________________

# 2.1. Big Words (BW)
Done using LIWC results. 

# t-test results
```{r}
# t-test of two independent samples (Democrats vs. Republcans)
t.test(liwc$Sixltr ~ liwc$Party.General, alternative = 'two.sided', conf.level = 0.95)

# Quality check - no need to report this
# QQ plots for Democrats and Republicans
qqnorm(liwc$Sixltr[liwc$Party.General == 'Democrat'], main = 'Democrats')
qqline(liwc$Sixltr[liwc$Party.General == 'Democrat'])

qqnorm(liwc$Sixltr[liwc$Party.General == 'Republican'], main = 'Republicans')
qqline(liwc$Sixltr[liwc$Party.General == 'Republican'])

# independent 2-group Mann-Whitney U Test 
wilcox.test(liwc$Sixltr ~ liwc$Party.General)
```

# Regression analysis

```{R}
# Multiple Linear Regression for Big Words

# Model 1 with full sample
fit1 <- lm(Sixltr ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = liwc[liwc$speechcount > 0, ])

# Model 2 with smaller sample based on speech count (more than 9 speeches given)
fit2 <- lm(Sixltr ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = liwc[liwc$speechcount > 9, ])

# Model 3 of interaction effects with party and sex with full sample 
fit3 <- lm(Sixltr ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + Party.General:Sex, # interaction term 
          data = liwc[liwc$speechcount > 0, ])

# Model 4 with addition of the statevote variable and a smaller sample
fit4 <- lm(Sixltr ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + statevote_2016, 
          data = liwc[liwc$speechcount > 0, ])

# show results
summary(fit1)
summary(fit2)
summary(fit3)
summary(fit4)
```

# Diagnostic checks of the regression models

```{R}
# Diagnostic plots (QQ, et al.)
layout(matrix(c(1,2,3,4), 2, 2))
plot(fit2)

# Boxplot
boxplot(Sixltr ~ Party.General,data = liwc[liwc$speechcount > 0,], 
        main="Sixltr", 
        xlab = "Party", ylab = "Big words (>6 letter)")
```

______________________________________________________________________________________________________

# 2.2. Words per sentence (WPS)
Done using LIWC results. 

# t-test results
```{r}
# t-test of two independent samples (Democrats vs. Republcans)
t.test(liwc$WPS ~ liwc$Party.General, alternative = 'two.sided', conf.level = 0.95)

# Quality check - no need to report this
# QQ plots for Democrats and Republicans
qqnorm(liwc$WPS[liwc$Party.General == 'Democrat'], main = 'Democrats')
qqline(liwc$WPS[liwc$Party.General == 'Democrat'])

qqnorm(liwc$WPS[liwc$Party.General == 'Republican'], main = 'Republicans')
qqline(liwc$WPS[liwc$Party.General == 'Republican'])

# independent 2-group Mann-Whitney U Test 
wilcox.test(liwc$WPS ~ liwc$Party.General)
```

# Regression analysis

```{R}
# Multiple Linear Regression for Big Words

# Model 1 with full sample
fit1 <- lm(WPS ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = liwc[liwc$speechcount > 0, ])

# Model 2 with smaller sample based on speech count (more than 9 speeches given)
fit2 <- lm(WPS ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = liwc[liwc$speechcount > 9, ])

# Model 3 of interaction effects with party and sex with full sample 
fit3 <- lm(WPS ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + Party.General:Sex, # interaction term 
          data = liwc[liwc$speechcount > 0, ])

# Model 4 with addition of the statevote variable and a smaller sample
fit4 <- lm(WPS ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + statevote_2016, 
          data = liwc[liwc$speechcount > 0, ])

# show results
summary(fit1)
summary(fit2)
summary(fit3)
summary(fit4)
```

# Diagnostic checks of the regression models

```{R}
# Diagnostic plots (QQ, et al.)
layout(matrix(c(1,2,3,4), 2, 2))
plot(fit2)

# Boxplot
boxplot(WPS ~ Party.General,data = liwc[liwc$speechcount > 0,], 
        main="WPS", 
        xlab = "Party", ylab = "Words per sentence")
```

______________________________________________________________________________________________________

# 3. Parts of Speech (POS)
______________________________________________________________________________________________________

POS tagging and creating variables for percentages of nouns and verbs per speaker.
```{r}
library("dplyr")
library("quanteda")
library("spacyr")
#spacy_initialize(model = "en")

-------------------------------------

  
-------------------------------------  
# create a corpus in quanteda 
pos_corpus <- corpus(pos_df)

# Create a list with the number of tokens per speaker
# to be added back in later
no_tokens <- ntoken(pos_corpus)

# Code to run the POS tagger on the LSD corpus
#pos_par <- spacy_parse(pos_corpus, tag = TRUE, pos = TRUE)

head(pos_par)

# Save created POS tagged file as an RDS file
# saveRDS(pos_par, "lsd_pos_tagged.rds")

# Load created POS tagged file
POS <- readRDS("lsd_pos_tagged.rds")

# create a dataframe with the variables ID and POS tags
df1 <- POS[, c("doc_id", "pos")]

# rename the ID variable so that it lines up with the other files
df1$ID <- as.integer(gsub('text', '', df1$doc_id))

# Create variables with total number of verbs, nouns, and proper nouns
df2 <- group_by(df1, ID) %>%
  summarise(count_verbs = sum(pos == 'VERB'), 
            count_nouns = sum(pos == 'NOUN'), 
            count_pnouns = sum(pos == 'PROPN'))

# add nouns and proper nouns in a combined variable
df2$count_nouns_total <- df2$count_nouns + df2$count_pnouns

# Add a variable for total number of tokens per speaker
df2$no_tokens <- no_tokens

# Calculate the percentage of POS per speaker
df2$verb <- df2$count_verbs/df2$no_tokens
df2$noun <- df2$count_nouns/df2$no_tokens
df2$noun_total <- df2$count_nouns_total/df2$no_tokens

# add new variables for percentages to LSD file
pos_df <- left_join(lsd, df2[,c('ID','verb','noun','noun_total')], by = 'ID')

# Convert to percentage
pos_df$verb <- round(100 * pos_df$verb, 2)
pos_df$noun <- round(100 * pos_df$noun, 2)
pos_df$noun_total <- round(100 * pos_df$noun_total, 2)

#save as RDS file
saveRDS(pos_df,"pos_results.rds")
```

# Boxplots
For verb and noun usage by party.

--> Should think about creating separate plots for verbs and noun_total. 

```{r}
library(ggthemes)
library(grid)
library(ggplot2)

# define fill and line colors
fill <- "#4271AE"
line <- "#1F3552"

# --------------------------

# Boxplot for both Positive and Negative emotion
df1 <- pos_df[,c('Party.General','verb')]
df2 <- pos_df[,c('Party.General','noun_total')]

colnames(df1) <- c('party','value')
colnames(df2) <- c('party','value')

df1$score <- 'Verbs'
df2$score <- 'Nouns'

# combine two dataframes
df <- rbind(df1, df2)

# define as factor
df$score <- factor(df$score)

# create plot - with different scales for verbs and nouns
p13 <- ggplot(df, aes(x = party, y = value)) +
        geom_boxplot(fill = fill, colour = line,
                     alpha = 0.7,
                     width = 0.6, position = position_dodge2(preserve = "total")) +
        scale_y_continuous(name = "Percentage of usage") +
        scale_x_discrete(name = "Party") +
        ggtitle("") +
        theme_bw() +
        theme(plot.title = element_text(size = 14, family = "Tahoma", face = "bold"),
              text = element_text(size = 12, family = "Tahoma"),
              axis.title = element_text(face="bold"),
              axis.text.x=element_text(size = 11)) +
        facet_grid(. ~ score, scales = "free") +
        coord_flip()

# create plot - equally scaled version
p11 <- ggplot(df, aes(x = party, y = value)) +
        geom_boxplot(fill = fill, colour = line,
                     alpha = 0.7,
                     width = 0.6, position = position_dodge2(preserve = "total")) +
        scale_y_continuous(name = "Percentage of noun (left) and verb (right) usage",
                           breaks = seq(8, 44, 2),
                           limits=c(8, 44)) +
        scale_x_discrete(name = "Party") +
        ggtitle("") +
        theme_bw() +
        theme(plot.title = element_text(size = 14, family = "Tahoma", face = "bold"),
              text = element_text(size = 12, family = "Tahoma"),
              axis.title = element_text(face="bold"),
              axis.text.x=element_text(size = 11)) +
        facet_grid(. ~ score) +
        coord_flip()                
      
        
```

Boxplot comparing nouns to nouns total
```{R}
# Boxplot for both Positive and Negative emotion
df1 <- pos_df[,c('Party.General','noun_total')]
df2 <- pos_df[,c('Party.General','noun')]

colnames(df1) <- c('party','value')
colnames(df2) <- c('party','value')

df1$score <- 'noun_total'
df2$score <- 'noun'

# combine two dataframes
df <- rbind(df1, df2)

# define as factor
df$score <- factor(df$score)

# create plot - with different scales for total nouns and nouns
p11 <- ggplot(df, aes(x = party, y = value)) +
        geom_boxplot(fill = fill, colour = line,
                     alpha = 0.7,
                     width = 0.6, position = position_dodge2(preserve = "total")) +
        scale_y_continuous(name = "Percentage of noun usage (left) and total noun usage (right)") +
        scale_x_discrete(name = "Party") +
        ggtitle("") +
        theme_bw() +
        theme(plot.title = element_text(size = 14, family = "Tahoma", face = "bold"),
              text = element_text(size = 12, family = "Tahoma"),
              axis.title = element_text(face="bold"),
              axis.text.x=element_text(size = 11)) +
        facet_grid(. ~ score, scales = "free") +
        coord_flip()

# with the same scales
p11 <- ggplot(df, aes(x = party, y = value)) +
        geom_boxplot(fill = fill, colour = line,
                     alpha = 0.7,
                     width = 0.6, position = position_dodge2(preserve = "total")) +
        scale_y_continuous(name = "Percentage of noun (left) and total noun usage (right)",
                           breaks = seq(10, 44, 5),
                           limits=c(10, 45)) +
        scale_x_discrete(name = "Party") +
        ggtitle("") +
        theme_bw() +
        theme(plot.title = element_text(size = 14, family = "Tahoma", face = "bold"),
              text = element_text(size = 12, family = "Tahoma"),
              axis.title = element_text(face="bold"),
              axis.text.x=element_text(size = 11)) +
        facet_grid(. ~ score) +
        coord_flip()  

```

______________________________________________________________________________________________________

# 3.1. Verb
Done using results from the POS tagging. 

# t-test results

```{r}
# --------------------------

# t-test of two independent samples (Democrats vs. Republcans)
t.test(pos_df$verb ~ pos_df$Party.General, alternative = 'two.sided', conf.level = 0.95)

# Quality check - no need to report this
# QQ plots for Democrats and Republicans
qqnorm(pos_df$verb[pos_df$Party.General == 'Democrat'], main = 'Democrats')
qqline(pos_df$verb[pos_df$Party.General == 'Democrat'])

qqnorm(pos_df$verb[pos_df$Party.General == 'Republican'], main = 'Republicans')
qqline(pos_df$verb[pos_df$Party.General == 'Republican'])

# independent 2-group Mann-Whitney U Test 
wilcox.test(pos_df$verb ~ pos_df$Party.General)
```

# Regression analysis

```{R}
# Multiple Linear Regression for Verb usage

# Model 1 with full sample
fit1 <- lm(verb ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = pos_df[pos_df$speechcount > 0, ])

# Model 2 with smaller sample based on speech count (more than 9 speeches given)
fit2 <- lm(verb ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = pos_df[pos_df$speechcount > 9, ])

# Model 3 of interaction effects with party and sex with full sample 
fit3 <- lm(verb ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + Party.General:Sex, # interaction term 
          data = pos_df[pos_df$speechcount > 0, ])

# Model 4 with addition of the statevote variable and a smaller sample
fit4 <- lm(verb ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + statevote_2016, 
          data = pos_df[pos_df$speechcount > 0, ])

# show results
summary(fit1)
summary(fit2)
summary(fit3)
summary(fit4)
```

# Diagnostic checks of the regression models

```{R}
# Diagnostic plots (QQ, et al.)
layout(matrix(c(1,2,3,4), 2, 2))
plot(fit2)

# Boxplot for verb usage
layout(matrix(c(1,1), 2, 2))
boxplot(verb ~ Party.General,data = pos_df[pos_df$speechcount > 9,], 
        main="Verb", 
        xlab = "Party", ylab = "Verb Usage")
```

______________________________________________________________________________________________________

# 3.2. Nouns
Using results from the POS tagging. 

# t-test results

```{r}
# t-test of two independent samples (Democrats vs. Republcans)
t.test(pos_df$noun ~ pos_df$Party.General, 
       alternative = 'two.sided', conf.level = 0.95)
# t-test of all nouns
t.test(pos_df$noun_total ~ pos_df$Party.General, 
       alternative = 'two.sided', conf.level = 0.95)

# Quality check - no need to report this
# QQ plots for Democrats and Republicans
qqnorm(pos_df$noun_total[pos_df$Party.General == 'Democrat'], main = 'Democrats')
qqline(pos_df$noun_total[pos_df$Party.General == 'Democrat'])

qqnorm(pos_df$noun_total[pos_df$Party.General == 'Republican'], main = 'Republicans')
qqline(pos_df$noun_total[pos_df$Party.General == 'Republican'])

# independent 2-group Mann-Whitney U Test 
wilcox.test(pos_df$noun ~ pos_df$Party.General)
# all nouns
wilcox.test(pos_df$noun_total ~ pos_df$Party.General)
```

# Regression analysis
Analysis for nouns
```{r}
# Model 1 with full sample
fit1 <- lm(noun ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = pos_df[pos_df$speechcount > 0, ])

# Model 2 with smaller sample based on speech count (more than 9 speeches given)
fit2 <- lm(noun ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = pos_df[pos_df$speechcount > 9, ])

# Model 3 of interaction effects with party and sex with full sample 
fit3 <- lm(noun ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + Party.General:Sex, # interaction term 
          data = pos_df[pos_df$speechcount > 0, ])

# Model 4 with addition of the statevote variable and a smaller sample
fit4 <- lm(noun ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + statevote_2016, 
          data = pos_df[pos_df$speechcount > 0, ])
```

Analysis for all nouns
```{R}
# Multiple Linear Regression for All Nouns

# Model 1 with full sample
fit1 <- lm(noun_total ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = pos_df[pos_df$speechcount > 0, ])

# Model 2 with smaller sample based on speech count (more than 9 speeches given)
fit2 <- lm(noun_total ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = pos_df[pos_df$speechcount > 9, ])

# Model 3 of interaction effects with party and sex with full sample 
fit3 <- lm(noun_total ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + Party.General:Sex, # interaction term 
          data = pos_df[pos_df$speechcount > 0, ])

# Model 4 with addition of the statevote variable and a smaller sample
fit4 <- lm(noun_total ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + statevote_2016, 
          data = pos_df[pos_df$speechcount > 0, ])

# show results
summary(fit1)
summary(fit2)
summary(fit3)
summary(fit4)
```

# Diagnostic checks of the regression models

```{R}
# Diagnostic plots (QQ, et al.)
layout(matrix(c(1,2,3,4), 2, 2))
plot(fit2)

# Boxplot for noun usage
layout(matrix(c(1,1), 2, 2))
boxplot(noun ~ Party.General,data = pos_df[pos_df$speechcount > 0,], 
        main="Noun Usage", 
        xlab = "Party", ylab = "Nouns")

# Boxplot for total noun usage
layout(matrix(c(1,1), 2, 2))
boxplot(noun_total ~ Party.General,data = pos_df[pos_df$speechcount > 0,], 
        main="Total Noun Usage", 
        xlab = "Party", ylab = "Nouns")
```

# 3.3. Nouns Total
Using results from the POS tagging. 

# t-test results

```{r}
# t-test of two independent samples (Democrats vs. Republcans)
t.test(pos_df$noun_total ~ pos_df$Party.General, 
       alternative = 'two.sided', conf.level = 0.95)
# t-test of all nouns
t.test(pos_df$noun_total ~ pos_df$Party.General, 
       alternative = 'two.sided', conf.level = 0.95)

# Quality check - no need to report this
# QQ plots for Democrats and Republicans
qqnorm(pos_df$noun_total[pos_df$Party.General == 'Democrat'], main = 'Democrats')
qqline(pos_df$noun_total[pos_df$Party.General == 'Democrat'])

qqnorm(pos_df$noun_total[pos_df$Party.General == 'Republican'], main = 'Republicans')
qqline(pos_df$noun_total[pos_df$Party.General == 'Republican'])

# independent 2-group Mann-Whitney U Test 
wilcox.test(pos_df$noun_total ~ pos_df$Party.General)

```

# Regression analysis
Analysis for nouns
```{r}
# Model 1 with full sample
fit1 <- lm(noun ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = pos_df[pos_df$speechcount > 0, ])

# Model 2 with smaller sample based on speech count (more than 9 speeches given)
fit2 <- lm(noun_total ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = pos_df[pos_df$speechcount > 9, ])

# Model 3 of interaction effects with party and sex with full sample 
fit3 <- lm(noun_total ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + Party.General:Sex, # interaction term 
          data = pos_df[pos_df$speechcount > 0, ])

# Model 4 with addition of the statevote variable and a smaller sample
fit4 <- lm(noun_total ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + statevote_2016, 
          data = pos_df[pos_df$speechcount > 0, ])
```

Analysis for all nouns
```{R}
# Multiple Linear Regression for All Nouns

# Model 1 with full sample
fit1 <- lm(noun_total ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = pos_df[pos_df$speechcount > 0, ])

# Model 2 with smaller sample based on speech count (more than 9 speeches given)
fit2 <- lm(noun_total ~ Party.General + Sex + age + educ + seniority + Region +
            Christian, 
          data = pos_df[pos_df$speechcount > 9, ])

# Model 3 of interaction effects with party and sex with full sample 
fit3 <- lm(noun_total ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + Party.General:Sex, # interaction term 
           data = pos_df[pos_df$speechcount > 0, ])

# Model 4 with addition of the statevote variable and a smaller sample
fit4 <- lm(noun_total ~ Party.General + Sex + age + educ + seniority + Region +
            Christian + statevote_2016, 
          data = pos_df[pos_df$speechcount > 0, ])

# show results
summary(fit1)
summary(fit2)
summary(fit3)
summary(fit4)
```

# Diagnostic checks of the regression models

```{R}
# Diagnostic plots (QQ, et al.)
layout(matrix(c(1,2,3,4), 2, 2))
plot(fit2)

# Boxplot for noun usage
layout(matrix(c(1,1), 2, 2))
boxplot(noun ~ Party.General,data = pos_df[pos_df$speechcount > 0,], 
        main="Noun Usage", 
        xlab = "Party", ylab = "Nouns")

# Boxplot for total noun usage
layout(matrix(c(1,1), 2, 2))
boxplot(noun_total ~ Party.General,data = pos_df[pos_df$speechcount > 0,], 
        main="Total Noun Usage", 
        xlab = "Party", ylab = "Nouns")

table(alltexts_clean_lsd$speechcount)

```


______________________________________________________________________________________________________

# 4. Visualizations
______________________________________________________________________________________________________

# Principal Component Analysis

Create combined file
```{r}
# remove NAs
na.omit(pos_df)

pos_df_clean <- pos_df[-c(377), ]
liwc_clean <- liwc[-c(377), ]

df_pca <- cbind(liwc_clean[,c("ID", "Name", "Sixltr", "WPS")], pos_df_clean[,c("noun_total", "verb", "positive", "negative")])

df_pca <- cbind(liwc_clean[liwc_clean$speechcount > 9, c("ID", "Name", "Sixltr", "WPS")], pos_df_clean[pos_df_clean$speechcount > 9, c("noun_total", "verb", "positive", "negative")])

```


```{R}
#install_github("vqv/ggbiplot")
library("ggbiplot")

PCA_df <- prcomp(df_pca[,colnames(df_pca) %in% c("positive", "negative", "WPS", "Sixltr", "noun_total", "verb")], center = TRUE, scale. = TRUE)

summary(PCA_df)

# Plot PCA
ggbiplot(PCA_df)

# basic plot
pcanames <- ggbiplot(PCA_df, labels = liwc_clean$Name)

# without lables
pca_plot <- ggbiplot(PCA_df, ellipse=TRUE, obs.scale = .5, var.scale = .5, groups = liwc_clean$Party.General[liwc_clean$speechcount > 9]) +
  scale_colour_manual(name = "Party", values = c("dark blue", "red")) +
  ggtitle("") +
  theme_minimal() +
  theme(legend.position = "bottom")

# with lables
pca_plot2 <- ggbiplot(PCA_df, ellipse=TRUE, obs.scale = .5, var.scale = .5, lables = liwc_clean$Name[liwc_clean$speechcount > 9], groups = liwc_clean$Party.General[liwc_clean$speechcount > 9]) +
  geom_text(aes(label = liwc_clean$Name[liwc_clean$speechcount > 9]),hjust=0, vjust=0, size = 3) +
  scale_colour_manual(name = "Party", values = c("dark blue", "red")) +
  ggtitle("") +
  theme_minimal() +
  theme(legend.position = "bottom")
  

# Should also include complexity in another version of this plot

```



_________________________________________________________________________________________________
_________________________________________________________________________________________________

# Principal Component Analysis V2
Include only Emotion and POS

Create combined file
```{r}
# remove NAs
na.omit(pos_df)

pos_df_clean <- pos_df[-c(377), ]
liwc_clean <- liwc[-c(377), ]

df_pca <- pos_df_clean[,c("ID", "Name", "noun_total", "verb", "positive", "negative")]

# Reduced sample size
df_pca <- pos_df_clean[pos_df_clean$speechcount > 9, c("ID", "Name", "noun_total", "verb", "positive", "negative")]

# Less reduced sample
df_pca <- pos_df_clean[pos_df_clean$speechcount > 5, c("ID", "Name", "noun_total", "verb", "positive", "negative")]

# Full sample
df_pca <- pos_df_clean[pos_df_clean$speechcount > 0, c("ID", "Name", "noun_total", "verb", "positive", "negative")]

```


```{R}
#install_github("vqv/ggbiplot")
library("ggbiplot")

PCA_df <- prcomp(df_pca[,colnames(df_pca) %in% c("positive", "negative", "noun_total", "verb")], center = TRUE, scale. = TRUE)

summary(PCA_df)

# Plot PCA
ggbiplot(PCA_df)

# basic plot
pcanames <- ggbiplot(PCA_df, labels = liwc_clean$Name)

# without lables
pca_plot <- ggbiplot(PCA_df, ellipse=TRUE, obs.scale = .5, var.scale = .5, groups = liwc_clean$Party.General[liwc_clean$speechcount > 0]) +
  scale_colour_manual(name = "Party", values = c("dark blue", "red")) +
  ggtitle("") +
  theme_minimal() +
  theme(legend.position = "bottom")

# with lables
pca_plot2 <- ggbiplot(PCA_df, ellipse=TRUE, obs.scale = .5, var.scale = .5, lables = liwc_clean$Name[liwc_clean$speechcount > 9], groups = liwc_clean$Party.General[liwc_clean$speechcount > 9]) +
  geom_text(aes(label = liwc_clean$Name[liwc_clean$speechcount > 9]),hjust=0, vjust=0, size = 3) +
  scale_colour_manual(name = "Party", values = c("dark blue", "red")) +
  ggtitle("") +
  theme_minimal() +
  theme(legend.position = "bottom")
  

# Should also include complexity in another version of this plot

```
